Separate the CodeMirror field from the language it was written for
`CodeEditor` was the console's only CodeMirror, so everything in it was written for one caller: the
JavaScript grammar, the completion over declared arguments, the scope hover and its key, and the two
sentences that report the code tool's character cap. The sample response wants the same field with a
different grammar, and #563 will want a third.

So what is generic is now `CodeMirrorField` — the view built once, the controlled document, the
accessibility attributes on the contenteditable, the placeholder, the character cap and the Escape
claim — and what is about JavaScript stayed in `CodeEditor`, which is now that shell's first caller.

Two seams carry the split, and each is a way to silently cost the operator their cursor and their
undo history.

The language arrives as `extensions`, in a compartment, reconfigured on the prop's identity. That is
the seam the code tool already depended on (renaming a declared argument reconfigures the completion
source under a mounted editor), generalized. Listing it as a lifecycle dependency of the build would
answer that rename by throwing the editor away mid-keystroke.

The cap travels as ONE prop holding the number and the two sentences, because a cap without words is
the failure the cap machinery exists to avoid: the change filter refuses a paste whole, and a refusal
nobody sees is a field that stopped accepting text for no reason. Only the NUMBER reaches the view,
because an object is a fresh identity on every render of the caller and the editor is built against
the cap. The words are read at render, so "Shorten the body" cannot greet an operator whose sample
response is too long.

The 90 tests of `code-editor-completions` and `editor-text-caps` are the fence for the move being
behaviour-preserving: they drive `CodeEditor` and not one of them changed. The three new tests drive
the seam itself, and mutation kills each: building against the cap object, or listing the extensions
as a build dependency, each rebuilds the view and loses the history.
